flask/requirements.txt
tensorflow==1.15
bert-tensorflow==1.0.1
requests
flask
flask-cors